requirements.txt
# Telegram Bot Framework with extensions for persistence
python-telegram-bot[ext]

# System and process monitoring
psutil

# Google Generative AI API
google-generativeai
